+++ Pipfile
@@ -13,6 +13,7 @@ isort = "*"
 mypy = "*"
 pytest = "*"
 tox = "*"
+twine = "*"
 
 [requires]
 python_version = "3.11"
