+++ src/builder/__init__.py
@@ -1,2 +1,1 @@
 from .builder import Builder
-from .template import Template
